add toaster on fomulaire + refacto toaster
