+++ src/client/containers/form/HomeForm.js
@@ -4,13 +4,28 @@ import {Input, Error, Button} from "../../components/block";
 import {maxLength15, minLength3, required} from "../../utils/formValidation"
 import {connect} from "../../actions/user"
 import { push } from 'connected-react-router'
+import {notify} from "../../utils/eventHandler";
+
+const handleClick = (form, e) => {
+
+  if (form && form.hasOwnProperty('syncErrors')) {
+    if (form.syncErrors.userName) {
+      e.preventDefault()
+      notify('login: '+form.syncErrors.userName, 'error')
+    }
+    if (form.syncErrors.gameName) {
+      e.preventDefault()
+      notify('room: '+form.syncErrors.gameName, 'error')
+    }
+  }
+}
 
 const onSubmit = (data, dispatch) => {
   dispatch(push(`/#${data.gameName}/${data.userName}`))
   dispatch(connect(data))
 }
 
-const HomeForm = ({ handleSubmit, error, submitting }) => (
+const HomeForm = ({ handleSubmit, error, submitting, props }) => (
 <form id={'homeForm'} onSubmit={handleSubmit(onSubmit)}>
   <Field
     placeholder='Your name'
@@ -24,7 +39,7 @@ const HomeForm = ({ handleSubmit, error, submitting }) => (
     component={Input}
     validate={[required, minLength3, maxLength15]}
   />
-  <Button type={'submit'} disabled={submitting} size={'large'} fullWidth to='/new-game'>Go</Button>
+  <Button onClick={handleClick.bind(this, props.homeForm)} type={'submit'} disabled={submitting} size={'large'} fullWidth to='/new-game'>Go</Button>
 </form>
 )
 export default reduxForm({
