+++ src/client/containers/view/Home.js
@@ -1,11 +1,12 @@
 import React from 'react'
 import { Link } from 'react-router-dom';
 
-import { Box, Card, LoadingContainer } from '../../components/block'
+import { Box, Card, LoadingContainer, Toaster } from '../../components/block'
 import HomeForm from '../form/HomeForm'
 import { connect } from 'react-redux'
 import {init} from "../../actions/user";
 import {emitGetGames} from "../../actions/games";
+import {notify} from "../../utils/eventHandler";
 
 class Home extends React.Component {
 
@@ -21,11 +22,34 @@ class Home extends React.Component {
     getGames()
   }
 
+  handleClick(form, e) {
+
+    if (form && form.hasOwnProperty('syncErrors')) {
+      if (form.syncErrors.userName) {
+        e.preventDefault()
+        notify('login: '+form.syncErrors.userName, 'error')
+      }
+    }
+  }
+
   render () {
-    const { userName, gamesList, isLoading } = this.props
+    const { userName, gamesList, isLoading, homeForm } = this.props
+
+    const renderList = gamesList && gamesList.map((game, i) => {
+      const {gameName} = game
+      return(
+        <li key={i}>
+          <Box fontSize={30}>
+            <Link to={`/#${gameName}/${userName}`} onClick={this.handleClick.bind(this, homeForm)}>
+              {gameName}
+            </Link>
+          </Box>
+        </li>)
+    })
 
     return (
       <Box width={'100%'} flex flexDirection='row' justifyContent='center'>
+        <Toaster/>
         <Box flex={1}>
           <HomeForm props={{...this.props}}/>
         </Box>
@@ -36,7 +60,7 @@ class Home extends React.Component {
             emptyLabel='Pas de parties en cours'
           >
           <Box fontSize={30}>
-            {userName && userName.length > 0 ? <ul>{gamesList}</ul> : <div>Choose a login first</div>}
+            {userName && userName.length > 0 ? <ul>{renderList}</ul> : <div>Choose a login first</div>}
           </Box>
           </LoadingContainer>
         </Card>
@@ -56,27 +80,16 @@ const mapStateToProps = state => {
   const gamesList = state.games.items
   const userName = [];
 
-  if (homeForm && homeForm.hasOwnProperty('values') && homeForm.hasOwnProperty('syncErrors')) {
-    if (homeForm.values.userName && !homeForm.syncErrors.userName) {
+  if (homeForm && homeForm.hasOwnProperty('values')) {
+    if (homeForm.values.userName) {
       userName.push(homeForm.values.userName);
     }
   }
 
-  const renderList = gamesList && gamesList.map((game, i) => {
-    const {gameName} = game
-    return(
-      <li key={i}>
-        <Box fontSize={30}>
-          <Link to={`/#${gameName}/${userName}`}>
-            {gameName}
-          </Link>
-        </Box>
-      </li>)
-  })
-
   return {
     userName: userName,
-    gamesList: renderList,
+    gamesList: gamesList,
+    homeForm: homeForm,
     isLoading: state.games.isLoading,
   }
 };
