+++ src/client/utils/eventHandler.js
@@ -1,5 +1,6 @@
 import {tetriAction} from "../actions/tetrimino";
 import {store} from "../index";
+import {toast} from "react-toastify";
 
 const eventHandler = (event) => {
   const user = store.getState().user;
@@ -10,4 +11,61 @@ const eventHandler = (event) => {
   }
 };
 
-export {eventHandler};
+const notify = (message, type) => {
+
+  switch (type) {
+    case 'success': {
+      toast.success(message, {
+        position: toast.POSITION.TOP_RIGHT,
+        autoClose: 5000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,
+      });
+      break;
+    }
+    case 'error': {
+      toast.error(message, {
+        position: toast.POSITION.TOP_RIGHT,
+        autoClose: 5000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,      });
+      break;
+    }
+    case 'warning': {
+      toast.warn(message, {
+        position: toast.POSITION.TOP_RIGHT,
+        autoClose: 5000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,      });
+      break;
+    }
+    case 'info': {
+      toast.info(message, {
+        position: toast.POSITION.TOP_RIGHT,
+        autoClose: 5000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,      });
+      break;
+    }
+    default: {
+      toast(message, {
+        position: toast.POSITION.TOP_RIGHT,
+        autoClose: 5000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,      });
+      break;
+    }
+  }
+};
+
+export {eventHandler, notify};
